requirements/local.txt
altgraph==0.17
click==7.1.2
future==0.18.2
pefile==2019.4.18
Pillow==8.0.1
pycryptodome==3.9.8
pyinstaller==4.0
pyinstaller-hooks-contrib==2020.8
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyqt5-tools==5.15.1.1.7.3
python-dotenv==0.14.0
pywin32-ctypes==0.2.0